밤부 스코어 오류 해결 및 리뷰 인서트 기능 추가
2023.04.24

+++ jPanda/src/main/java/com/kakao/jPanda/lhw/domain/Review.java
@@ -14,6 +14,6 @@ public class Review {
 	private String reviewerId;
 	private String content;
 	private Date reviewDate;
-	private Long bambooScore;
+	private Double bambooScore;
 
 }
